Make write batch tests more consistent
Fix the kv_op_direct_addr helper to be more consistent with the existing helpers, and refactor the  C++ putAddr() methods to use that.

Repair some size flaws in the write batch tests; we were measuring for tiny writes, we should consistently measure for writes of known sizes; this makes the performance gain of the new methods less enormous, though still very significant.