Speculative critical section Get()
Refactor GetBenchmarks to have per-thread state, allow running multi-threaded.
Build cut'n'paste GetCritical methods to test use of critical section for copying back results, instead of SetByteArrayRegion().
I do not expect them to be faster, but it is possible.
I don’t think that critical sections offer much other performance possibility the way the code is now (JDirectBufferSlice and JByteArraySlice are as good as possible),
except maybe for large-sized keys, as keys are being copied and we could save this ?