All variants of JNI API putCritical
And tests for them. The tests are cut and paste from ColumnFamilyTest and we should really share them. An abstract ColumnFamilyTest would do it.. But we want to decide whether critical section is a good idea at all, and we should just replace it or not replace it in vanilla Java API put() methods..