Experiment with RocksDB object owning iterators
In an analogous way to owning column family handles, and closing them when the DB itself is closed, we make iterators be “child” objects of the DB.
Iterator lifecycle is generally shorter than DB, so we give them a lambda to remove themselves from the DB on iterator close().
Add a few tests to check that iterators which remain open when the DB is requested to close, are now closed when the DB is closed.

TODO - other DB subclasses need to add iterators to the list
TODO - there are cases where isOwningHandle() should not throw an exception; we need a less sledgehammer way of implementing this
TODO - fix the current Java tests
TODO - the atomic improvement in https://github.com/facebook/rocksdb/issues/5234