Experimental Java buffer of put()s in write batch
Minimal implementation which allows a write batch which only contains put()s to be buffered on the Java side, and applied once (over JNI) to create the C++ write batch
when the batch is flush()ed - flush() is a new operation.

Using any other operations other than put() will either crash, or not work as you might expect.