Block DB from close() until iterators are closed
Prevent SEGFAULTs on Java access to live iterators which are part of a closed DB.
The correct pattern of use for iterators is to nest their lifetime within that of the DB,
open DB
  open iterator
    use iterator
  close iterator
close DB
- but this pattern is not always respected. When it is not, if access is continued via
the iterator within a closed DB, Rocks may crash in C++, as it is accessing deleted
structures.
We introduce some co-ordination so that
1. The iterators opened within a DB are recorded
2. Closing a DB blocks the creation of further iterators
3. Closing a DB is blocked until all outstanding iterators are themselves closed.
This may result in existing code which is non-compliant, but which due to luck does not crash, to block instead. If iterators are not closed but are simply abandoned, closing the DB will hang permanently.