test prefetch optimization for RocksIterator
Enough prefetch (on isValid() and on seekToFirst()) to allow batch performance for sequential scan using iterator.

Passes RocksIteratorTests, but I know there are cases where it is broken.
TODO fix these cases (there is a TODO in isValid()
TODO handle the legacy case where we don’t prefetch to make it the default,
unless the caller uses newIterator(<prefetch-size>)